tests/wpt/tests/js-self-profiling/js-profiling-mode/eager.https.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
  <script>
    // These tests verify that js-profiling-mode=eager is recognised as a valid
    // enum value and that profiling is allowed. The observable JS behaviour is
    // identical to lazy mode — the difference (V8 profiler initialised with
    // eager JIT logging at context-add time rather than deferred) is internal
    // and not detectable from the JS API.
    promise_test(async t => {
      const profiler = new Profiler({
        sampleInterval: 10,
        maxBufferSize: Number.MAX_SAFE_INTEGER,
      });
      assert_false(profiler.stopped, 'profiler should be running');
      await profiler.stop();
    }, 'js-profiling-mode=eager should allow profiling');

    promise_test(async t => {
      const profiler_a = new Profiler({
        sampleInterval: 10,
        maxBufferSize: Number.MAX_SAFE_INTEGER,
      });
      const profiler_b = new Profiler({
        sampleInterval: 10,
        maxBufferSize: Number.MAX_SAFE_INTEGER,
      });
      await profiler_b.stop();
      await profiler_a.stop();
    }, 'js-profiling-mode=eager should support concurrent profilers');
  </script>
</body>
</html>
